Refactor UrlParser into RepoUrl service
