+++ application/test/services/repo/repo_url_test.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 require 'test_helper'
 
-class UrlParserTest < ActiveSupport::TestCase
+class Repo::RepoUrlTest < ActiveSupport::TestCase
   test 'should parse standard HTTPS URL' do
     url = 'https://example.com/path/to/resource?foo=bar&baz=qux'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     assert parser.https?
@@ -18,7 +18,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should parse standard HTTP URL' do
     url = 'http://example.org/foo/bar?x=1&y=2'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     refute parser.https?
@@ -32,7 +32,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should parse HTTP URL with custom port' do
     url = 'http://localhost:3001/custom/path?one=1'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     refute parser.https?
@@ -46,7 +46,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should handle URL with no query params' do
     url = 'https://example.com/simple/path'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     assert_equal '/simple/path', parser.path
@@ -55,7 +55,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should handle URL with no path' do
     url = 'https://example.com'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     assert_equal 'https', parser.scheme
@@ -68,7 +68,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should handle URL with domain and no protocol' do
     url = 'www.example.com'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     assert_equal 'https', parser.scheme
@@ -80,17 +80,17 @@ class UrlParserTest < ActiveSupport::TestCase
   end
 
   test 'path should default to /' do
-    assert_equal '/', UrlParser.parse('https://example.com').path
-    assert_equal [], UrlParser.parse('https://example.com').path_segments
-    assert_equal '/', UrlParser.parse('https://example.com/').path
-    assert_equal [], UrlParser.parse('https://example.com/').path_segments
-    assert_equal '/', UrlParser.parse('https://example.com//').path
-    assert_equal [], UrlParser.parse('https://example.com//').path_segments
+    assert_equal '/', Repo::RepoUrl.parse('https://example.com').path
+    assert_equal [], Repo::RepoUrl.parse('https://example.com').path_segments
+    assert_equal '/', Repo::RepoUrl.parse('https://example.com/').path
+    assert_equal [], Repo::RepoUrl.parse('https://example.com/').path_segments
+    assert_equal '/', Repo::RepoUrl.parse('https://example.com//').path
+    assert_equal [], Repo::RepoUrl.parse('https://example.com//').path_segments
   end
 
   test 'should normalize path with multiple slashes' do
     url = 'https://example.com//test/value//other'
-    parser = UrlParser.parse(url)
+    parser = Repo::RepoUrl.parse(url)
 
     assert parser
     assert_equal '/test/value/other', parser.path
@@ -98,16 +98,16 @@ class UrlParserTest < ActiveSupport::TestCase
   end
 
   test 'should return nil for nil URL' do
-    assert_nil UrlParser.parse(nil)
+    assert_nil Repo::RepoUrl.parse(nil)
   end
 
   test 'should return nil for invalid URL' do
-    assert_nil UrlParser.parse('not a url')
-    assert_nil UrlParser.parse('123://bad')
+    assert_nil Repo::RepoUrl.parse('not a url')
+    assert_nil Repo::RepoUrl.parse('123://bad')
   end
 
   test 'build should default to https scheme' do
-    parser = UrlParser.build('example.org')
+    parser = Repo::RepoUrl.build('example.org')
 
     assert parser
     assert_equal 'https', parser.scheme
@@ -116,7 +116,7 @@ class UrlParserTest < ActiveSupport::TestCase
   end
 
   test 'build should accept custom scheme and port' do
-    parser = UrlParser.build('localhost', scheme: 'http', port: 8080)
+    parser = Repo::RepoUrl.build('localhost', scheme: 'http', port: 8080)
 
     assert parser
     assert_equal 'http', parser.scheme
@@ -125,22 +125,22 @@ class UrlParserTest < ActiveSupport::TestCase
   end
 
   test 'build should return nil when domain is blank' do
-    assert_nil UrlParser.build(nil)
-    assert_nil UrlParser.build('')
+    assert_nil Repo::RepoUrl.build(nil)
+    assert_nil Repo::RepoUrl.build('')
   end
 
   test 'scheme_override returns nil for https and scheme for others' do
-    https_parser = UrlParser.parse('https://example.com')
-    http_parser = UrlParser.parse('http://example.com')
+    https_parser = Repo::RepoUrl.parse('https://example.com')
+    http_parser = Repo::RepoUrl.parse('http://example.com')
 
     assert_nil https_parser.scheme_override
     assert_equal 'http', http_parser.scheme_override
   end
 
   test 'port_override returns nil for default ports and value for non-standard ports' do
-    https_default = UrlParser.parse('https://example.com')
-    http_default = UrlParser.parse('http://example.com:80')
-    http_custom = UrlParser.parse('http://example.com:8080')
+    https_default = Repo::RepoUrl.parse('https://example.com')
+    http_default = Repo::RepoUrl.parse('http://example.com:80')
+    http_custom = Repo::RepoUrl.parse('http://example.com:8080')
 
     assert_nil https_default.port_override
     assert_nil http_default.port_override
@@ -149,7 +149,7 @@ class UrlParserTest < ActiveSupport::TestCase
 
   test 'should raise NoMethodError when calling new directly' do
     assert_raises(NoMethodError) do
-      UrlParser.new('https://example.com/path')
+      Repo::RepoUrl.new('https://example.com/path')
     end
   end
 end
